pyproject.toml
[tool.poetry]
name = "if-heatmap"
version = "0.1.0"
description = ""
authors = ["Ismael <ismaelhenriqueeu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.5"
streamlit = "^1.38.0"
pandas = "^2.2.2"
pydantic = "^2.8.2"
ruff = "^0.6.3"
pre-commit = "^3.8.0"
python-dotenv = "^1.0.1"
sentry-sdk = "^2.13.0"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.2"
folium = "^0.17.0"
flask = "^3.0.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
